Avoid unselecting the base product (bsc#1165501)
During debugging it turned out that the base product (like SLES)
was deselected at this point. Because it is later again selected by solver
by other modules dependencies it did not make any problems. But lets fix
it to be on the safe side.

In an Offline installation this is called at a different place and does
not make troubles so handle only the Online case.

- 4.2.58