Do not add an empty repository to the system (bsc#1204399)
When using Full medium in registered system the request
is to behave as with the Online medium. So we explicitly
override the detected medium type in that case.

But in that case an empty repository from the Full medium
root was added to the system because the medium type
condition was not met.

The solution is to remember the original detected medium
type and use it in this test.